# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: XiaoTian Quant Gateway API
  version: 1.0.0
  description: 由 gateway/cmd/apidump 自动生成的全路由清单契约（v1）：每条路由含 method+path+handler+鉴权标注，request/response schema 后续逐步补充。请勿手改；重新生成：cd gateway && go run ./cmd/apidump -openapi ../docs/api/openapi.yaml
servers:
  - url: /
    description: 网关根：REST 挂 /api，WebSocket 挂 /ws、/ws/v2
...
paths:
...
  /api/agent/cron:
    get:
      operationId: handler.AgentCronList
      summary: handler.AgentCronList
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentCronList
      x-source: cmd/server/router.go:714
    post:
      operationId: handler.AgentCronCreate
      summary: handler.AgentCronCreate
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentCronCreate
      x-source: cmd/server/router.go:715
  /api/agent/cron/{id}:
    delete:
      operationId: handler.AgentCronDelete
      summary: handler.AgentCronDelete
      tags:
        - agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentCronDelete
      x-source: cmd/server/router.go:717
  /api/agent/cron/{id}/run:
    post:
      operationId: handler.AgentCronRun
      summary: handler.AgentCronRun
      tags:
        - agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentCronRun
      x-source: cmd/server/router.go:718
  /api/agent/cron/{id}/toggle:
    post:
      operationId: handler.AgentCronToggle
      summary: handler.AgentCronToggle
      tags:
        - agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentCronToggle
      x-source: cmd/server/router.go:716
  /api/agent/memory:
    get:
      operationId: handler.AgentMemoryList
      summary: handler.AgentMemoryList
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentMemoryList
      x-source: cmd/server/router.go:719
    post:
      operationId: handler.AgentMemoryCreate
      summary: handler.AgentMemoryCreate
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentMemoryCreate
      x-source: cmd/server/router.go:720
  /api/agent/memory/{id}:
    delete:
      operationId: handler.AgentMemoryDelete
      summary: handler.AgentMemoryDelete
      tags:
        - agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentMemoryDelete
      x-source: cmd/server/router.go:721
...
  /api/agent/skills:
    get:
      operationId: handler.AgentSkillsList
      summary: handler.AgentSkillsList
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentSkillsList
      x-source: cmd/server/router.go:722
    post:
      operationId: handler.AgentSkillCreate
      summary: handler.AgentSkillCreate
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentSkillCreate
      x-source: cmd/server/router.go:723
  /api/agent/skills/{id}:
    delete:
      operationId: handler.AgentSkillDelete
      summary: handler.AgentSkillDelete
      tags:
        - agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentSkillDelete
      x-source: cmd/server/router.go:724
  /api/agent/telegram/pair-code:
    post:
      operationId: handler.AgentTelegramPairCode
      summary: handler.AgentTelegramPairCode
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentTelegramPairCode
      x-source: cmd/server/router.go:726
  /api/agent/telegram/status:
    get:
      operationId: handler.AgentTelegramStatus
      summary: handler.AgentTelegramStatus
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentTelegramStatus
      x-source: cmd/server/router.go:725
  /api/agent/telegram/unlink:
    post:
      operationId: handler.AgentTelegramUnlink
      summary: handler.AgentTelegramUnlink
      tags:
        - agent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功响应（v1 契约未细化 schema，见 x-handler 对应实现）
        "401":
          description: 未认证或 token 失效
      x-handler: handler.AgentTelegramUnlink
      x-source: cmd/server/router.go:727
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '登录态 JWT（Authorization: Bearer <token>），由 middleware.AuthRequired 校验'
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe 服务端回调验签头（v1 scheme）
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: TradingView/通用 webhook 的 HMAC-SHA256 签名（WEBHOOK_SECRET），验签在 handler 内完成